robotics/pyproject.toml
[tool.poetry]
name = "robotics"
version = "0.1.0"
description = ""
authors = ["Bart Timmermans <bti.timmermans@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22.4"
opencv-python = "^4.6.0"
gym = "0.21"
pyserial = "^3.5"
stable-baselines3 = "^1.5.0"
tensorboard = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
